Extract the Readline carrier into its own module.
